Complete code rewrite and addition of Max patch/standalone app.
Although this is a "fork" of Qdot's Keepoff project, the MyKeepon.ino
Arduino code is a complete rewrite with a fully-featured interface for
sending commands to My Keepon (and receiving polled data) using the
simple strings described in the README (once the Arduino has been
connected to the toy's I2C bus using a simple 4-wire ribbon cable).
We've also provided a Max patch and standalone app that can be used to
teleoperate My Keepon.